refactor!: remove optional decorator from id property of DTOs
The "id" property of a DTO will always be defined, when receiving data from the API.
When creating a DTO in a client, the "id" might need to be set to null or undefined explicitly now to avoid type errors.
Generally, predefined "id"s will be ignored by the API, unless explicitly allowed (i.e. course creation with custom id).

This might cause breaking changes in generated clients that generate different types for nullable properties.